backend/requirements.txt
asgiref==3.8.1
astroid==3.3.8
attrs==25.1.0
black==25.1.0
click==8.1.8
colorama==0.4.6
dill==0.3.9
Django==5.1.6
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
